replay: move tile context to separate header
Fixes a log bug, where the replay tile's C path is

    WARNING 04-23 17:37:06.458616 2029868 15   replay:0 src/app/firedancer-dev/commands/../../../discof/replay/fd_replay_tile.c(2295): banks are full and partially executed frontier banks are being evicted

instead of

    WARNING 04-23 17:37:06.458616 2029868 15   replay:0 src/discof/replay/fd_replay_tile.c(2295): banks are full and partially executed frontier banks are being evicted
